update profile + auth => project completed and ready for review

+++ app/controllers/TodoController.php
@@ -5,7 +5,7 @@ namespace app\controllers;
 use app\models\Todo;
 use app\views\Viewer;
 
-class TodoController
+class TodoController extends AuthController
 {
     public $model;
     public function __construct() 
@@ -15,6 +15,8 @@ class TodoController
 
     public function index()
     {
+        $this->authorize();
+
         $todos = $this->model->getTodos();
         $viewer = new Viewer();
         $viewer->render("todos/index.php" , ['todos' => $todos]);
@@ -22,6 +24,8 @@ class TodoController
     }
     public function create()
     {
+        $this->authorize();
+
         $viewer = new Viewer();
         $viewer->render("todos/create.php");
     }
@@ -29,6 +33,7 @@ class TodoController
     public function store($data)
     {
 
+        $this->authorize();
         
         $errors = [];
         $text = trim($data['text']);
@@ -50,6 +55,8 @@ class TodoController
 
     public function edit($id)
     {
+        $this->authorize();
+
         $todo = $this->model->getTodosById($id);
         $viewer = new Viewer();
         $viewer->render("todos/edit.php" , ['todo' => $todo]);
@@ -58,6 +65,8 @@ class TodoController
 
     public function update($data , $id) 
     {
+        $this->authorize();
+
         $errors = [];
         $text = trim($data['text']);
         
@@ -79,6 +88,7 @@ class TodoController
 
     public function delete($data , $id)
     {
+        $this->authorize();
 
         $this->model->deleteTodo($id) ;
         redirect("/todos/index");
